Noaax v0.1.0 Noaax.ListStations

List station from a given state

Functions

code_name(list)

Receive a list, returns a list of tuples [{station_code, station_name}]

Example

iex> list = [
...>          {"td", [{"headers", "Station Name"}],
...>            [{"a", [{"href", "display.php?stid=KAAT"}], ["Alturas"]}, " (KAAT)"]},
...>          {"td", [{"headers", "Station Name"}],
...>            [{"a", [{"href", "display.php?stid=KACV"}], ["Arcata Airport"]}, " (KACV)"]}
...>        ]
iex> Noaax.ListStations.code_name(list)
[{"KAAT", "Alturas"}, {"KACV", "Arcata Airport"}]

delete_parenthesis(str)

Receive a string with parenthesis, returns a string without

Examples

iex> Noaax.ListStations.delete_parenthesis("(KJNU)")
"KJNU"
iex> Noaax.ListStations.delete_parenthesis("L56B")
"L56B"

print(data)