Noaax v0.1.0 Noaax.NoaaService

Fetch NOAA web page

Functions

data_station_url(station)

http://w1.weather.gov/xml/current_obs/STATION.xml

fetch(station)

Fetch station data, will return a tuple in the form of { :ok, body } or { :error, body }

fetch_list_station(statename)

Fetch station data, will return a tuple in the form of { :ok, body } or { :error, body }

handle_response(arg)

Return {:ok, body} if http status code is 200

Return {:error, body} if http status code is not 200

stations_url(statename)

http://w2.weather.gov/xml/current_obs/seek.php?state=STATENAME&Find=Find